[logger] (csv aggregation) move to dedicated file
