+++ tricicl/research/mathieu/display_aggregated_mnist_results.py
@@ -1,43 +1,6 @@
-from typing import List, Tuple
-
 import seaborn
-from matplotlib.figure import Figure
-from pandas import read_csv
-from seaborn import FacetGrid, relplot
 
-from tricicl.research.mathieu.aggregate_split_mnist_results import get_aggregate_csv_file
-
-
-def display_aggregated_results(
-    task_name: str, *, use_simplified_metric_name: bool = False, metrics_names: List[str] = None
-):
-    df = read_csv(get_aggregate_csv_file(task_name))
-
-    if use_simplified_metric_name:
-        df["metric"] = df["metric"].map(lambda s: s.replace("/eval_phase/test_stream", "").replace("/Task000", ""))
-
-    metrics_names = metrics_names or sorted(set(df["metric"]), key=get_metric_name_priority)
-
-    g: FacetGrid = relplot(
-        data=df,
-        kind="line",
-        x="step",
-        y="value",
-        hue="run_algo",
-        col="metric",
-        col_order=metrics_names,
-        col_wrap=min(3, len(metrics_names)),
-        facet_kws={"sharex": False, "legend_out": False},
-    )
-    fig: Figure = g.fig
-    fig.suptitle(task_name, fontsize=16)
-    fig.tight_layout()
-    fig.show()
-
-
-def get_metric_name_priority(name: str) -> Tuple[bool, str]:
-    return name[-1].isnumeric(), name
-
+from tricicl.loggers.csv_aggregation import display_aggregated_results
 
 if __name__ == "__main__":
     seaborn.set()
